Move origin storage type to `server_structs` package
